bpf: Disallow attaching same prog to cgroup through new links
When looking for a cgroup attach entry to slot a bpf prog, we make sure
to not attach a redundant program. With the introduction of bpf links,
we also added a check to make sure we don't attach the same link twice.
Yet, we still allow a redundant prog to be attached through mixed attach
style (raw attach alongside a link attach) or through creating new links
representing the same prog attach. This patch fixes that inconsistency.

Fixes: af6eea57437a ("bpf: Implement bpf_link-based cgroup BPF program attachment")
Signed-off-by: Milan Landaverde <milan@mdaverde.com>